// src/shared/dataprovider/api/model/eODVendorSummaryReport.ts
/**
 * DynastyGateway API
 * API Connections to DPay
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EODTopCards } from './eODTopCards';
import { VendorGrandTotals } from './vendorGrandTotals';
import { VendorReport } from './vendorReport';

export interface EODVendorSummaryReport { 
    EODTopCards?: EODTopCards;
    GrandTotals?: VendorGrandTotals;
    Reports?: Array<VendorReport>;
    TotalRecordCount?: number;
    Pagesize?: number;
    PageNumber?: number;
}